Borrado masivo por pasillo y fix de cantidades decimales en consultas
Borrado por pasillo
-------------------
Borrar un sector generado con parametros equivocados requeria eliminar
posicion por posicion (26-30 clics). Se agrega DELETE /locations/aisle
(ADMIN/MANAGER) y el boton "Eliminar pasillo" en el mapa de Depositos.

Es todo-o-nada: si alguna posicion tiene stock o pallets no borra ninguna
y devuelve cuales son. Un borrado parcial dejaria media estructura en pie,
mas dificil de entender y deshacer que un rechazo claro.

Cantidades decimales en consultas SQL
-------------------------------------
La migracion a numeric(14,3) dejo 9 consultas casteando cantidades a
::int, que ahora redondean:

- locations.service: la guarda de borrado reportaba stock 0 para una
  ubicacion con 0.4 unidades y la dejaba eliminar, dejando ese stock
  huerfano apuntando a un id inexistente. Es el caso grave.
- reports / warehouses / products: stock por producto, totales por
  deposito y reportes mostraban cifras redondeadas.

Se cambian a ::float8 las que son cantidades; los conteos siguen ::int.

Ademas: el ejemplo del generador de estructura mostraba "R1" cuando la
nomenclatura ya genera "F1".

+++ logistica-palets-backend/src/modules/products/products.service.ts
@@ -116,7 +116,7 @@ export class ProductsService {
          (SELECT COUNT(*) FROM lots WHERE "productId" = $1)::int              AS lots,
          (SELECT COUNT(*) FROM movements WHERE "productId" = $1)::int         AS movements,
          (SELECT COALESCE(SUM("currentQuantity"), 0) FROM stocks
-           WHERE "productId" = $1)::int                                       AS stock`,
+           WHERE "productId" = $1)::float8                                    AS stock`,
       [id],
     )) as Array<{ lots: number; movements: number; stock: number }>;
